Replace project.clj with deps.edn
Before the PRs adding support for async middleware
and the status-to-level configuration are merged to master,
I have to use my own fork as a git dependency. To do that, the library
must use deps, not Leiningen. This deps.edn contains only the bare minimum
dependencies, it doesn't even contain a way to run tests.

When the aforementioned PRs are merged, this branch should not be deleted.